requirements.txt
requests==2.32.3
kafka-python==2.0.2
pymongo==4.15.5
python-dotenv==1.2.1

pytest==9.0.2
pytest-mock==3.15.1

six==1.17.0
